scripts/requirements.txt
google-generativeai>=0.3.2